requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
asyncpg==0.27.0
authlib==1.3.0
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
pytest-cov==4.1.0
async-timeout==4.0.3
aiosqlite==0.21.0 